fix: change AssignorStaffID/AssigneeStaffID from Long to Integer
DB column type is int (INTEGER), not bigint — Hibernate schema validation
was failing with 'found int but expecting bigint'. Changed entity, DTO,
repository and service to use Integer consistently. Staff.Code (PK) remains
Long (bigint) since m03Staff.Code is bigint.

+++ src/main/java/com/aisolutions/jobtaskmanagement/dto/JobTaskDTO.java
@@ -55,8 +55,8 @@ public class JobTaskDTO {
         private String        taskTitle;
         private String        taskType;
         private String        taskDescription;
-        private Long          assignorStaffId;
-        private Long          assigneeStaffId;
+        private Integer      assignorStaffId;
+        private Integer      assigneeStaffId;
         private String        priority;
         private LocalDateTime dueDate;
         private BigDecimal    estimatedHours;
@@ -70,7 +70,7 @@ public class JobTaskDTO {
         private String        taskTitle;
         private String        taskType;
         private String        taskDescription;
-        private Long          assigneeStaffId;
+        private Integer      assigneeStaffId;
         private String        priority;
         private LocalDateTime dueDate;
         private BigDecimal    estimatedHours;
